evals/requirements.txt
anthropic>=0.40.0
pyyaml>=6.0
